ride_participantのモデルとコントローラを分離、リレーションの定義
