+++ backend/routes/web.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Api\Ride\MeetingPlaceController;
 use App\Http\Controllers\Api\Ride\RideRouteController;
 use App\Http\Controllers\Api\Ride\RideController;
 use App\Http\Controllers\RideViewController;
+use App\Http\Controllers\ParticipationController;
 
 /*
 |--------------------------------------------------------------------------
@@ -74,7 +75,7 @@ Route::group(['middleware' => ['auth']], function () {
 
     Route::post('api/post/createRide', [RideController::class, 'createRide'])->name('createRide');
 
-    Route::post('api/post/participation', [RideController::class, 'participationRegister'])->name('participationRegister');
+    Route::post('api/post/participation', [ParticipationController::class, 'participationRegister'])->name('participationRegister');
 
 });
